updating report
---

 1. Submission Files: All files included have been pushed into separete repo. report.html has been created, writeup_template.md has been filled. The python notebook has the necessary changes
 2. Dataset summary: The dataset summary such as number of training and test examples has been printed in the ipython notebook.
 3. Exploratory Visualization: All the classes images have been plotted to give an idea of the dataser. Also, a graph showing number of images in each class has also been plotted for the training dataset
 4. Preprocessing, Model architecture, model training, solution approach : Explained below
 5. Acquiring New Images: New images were downloaded. The images are placesd in a folder named 'custom_test_signs'
 6. Performance on new images: I tested on 5 new images and 4 of them were classified correctly, which gives an accuracy of 80%.
 7. Model Certainty - Softmax Probabilities: The softmax probabilities have been printed in the notebook

You're reading it! and here is a link to my [project code](https://github.com/vigneshu/traffic-sign-classifier/blob/master/Traffic_Sign_Classifier.ipynb)

I used the numpy library to calculate summary statistics of the traffic
signs data set:

* The size of training set is 47429
* The size of test set is 4410
* The shape of a traffic sign image is (32, 32, 3)
* The number of unique classes/labels in the data set is 43

Here is an exploratory visualization of the dataset.  It iterates through each class and finds the first image in the dataset which matches the class and displays it

![alt text][image1]

Next, I use the y_train to get the class count and plot a bar graph. You can see this in cell 4.

![alt text][image1]

I chose not to change it to grayscale because I felt the color gives good information for the network to classify the sign.
As a first step, I also created more data from the training datasset. The dataset does not have equal images for all classes, but rather skewed towards certain classes. This means that features for classes with less images aren't learn't and the model may be biased with the classes with lots of data. To overcome this, I took the classes with less images and randomly rotated them and added it to the training data again.

Finally, I normalised the image because it helps speed up learning and gives better results.

![alt text][image3]

The difference between the original data set and the augmented data set is the following ...

My final model consisted of the following layers:

| Layer         		|     Description	        					                                  |
|:---------------------:|:-------------------------------------------------------------------------------:|
| Input         		| Input = 32x32x3                                                                 |
| Convolution 3x3     	| Convolutional. Input = 32x32x3. Output = 28x28x6. 1x1 stride, 'VALID' padding   |
| RELU					|												                                  |
| Max pooling	      	| 2x2 stride,  outputs 14x14x6. 2x2 stride. 2x2 k-size                            |
| Convolution 3x3	    |  Convolutional. Input = 32x32x3. Output = 28x28x6. 1x1 stride, 'VALID' padding  |     				| RELU					|												                                  |
| Max pooling	      	| 2x2 stride,  Input = 10x10x16. Output = 5x5x16. 2x2 stride. 2x2 k-size	      |
| Fully connected		| Fully Connected. Input = 400. Output = 120. dropout = 0.3 (70% keep)            |
| RELU					|												                                  |
| Fully connected		| Fully Connected. Input = 84. Output = 10. dropout = 0.3 (70% keep)              |

To train the model, I used an LeNet. I added a dropout to the given Lenet model. The learning rate used was 0.001 and a batch size of 128. I used 10 EPOCHS and optimizer used is AdamOptimizer.

Training adn accuracy: cell 12

My final model results were:
* training set accuracy of 98.5
* validation set accuracy of  95.7
* test set accuracy of 95.3

If an iterative approach was chosen:
* What was the first architecture that was tried and why was it chosen?
THe first architecture was basically the ones provided by the instructors. I felt this would be a good starting point to get started with as LeNet is popularly used and gives good results.
* What were some problems with the initial architecture?
The main problem was the lack of data for certain classes. To overcome this a few images were rotated and added bacak to improve training.
* How was the architecture adjusted and why was it adjusted? Typical adjustments could include choosing a different model architecture, adding or taking away layers (pooling, dropout, convolution, etc), using an activation function or changing the activation function. One common justification for adjusting an architecture would be due to overfitting or underfitting. A high accuracy on the training set but low accuracy on the validation set indicates over fitting; a low accuracy on both sets indicates under fitting.
As mentioned above, Adding a dropout helped improving the training. Dropout prevents overfitting. WIthout the dropout the training gave in good results but the validation accuracy was low. This hinted overfitting and hence I added a dropout layer.
* Which parameters were tuned? How were they adjusted and why?
I tried out various EPOCH, BATCH size and learning rate. Increasing the EPOCH led to overfitting and therefore I stuck with the parameters given. I did not tune the other parameters also since it gave the required accuracy ayways.
* What are some of the important design choices and why were they chosen? For example, why might a convolution layer work well with this problem? How might a dropout layer help with creating a successful model?

If a well known architecture was chosen:
* What architecture was chosen?
LeNet architecture was used.
* Why did you believe it would be relevant to the traffic sign application?

* How does the final model's accuracy on the training, validation and test set provide evidence that the model is working well?
 The accuracy was above 93% in the three datasets which suggests the model works well. The model had never seen the validation and test data sets but still produced an excelklent efficiency which suggests a robust model.

Here are five German traffic signs that I found on the web:

![alt text][image4] ![alt text][image5] ![alt text][image6]
![alt text][image7] ![alt text][image8]

Here are the results of the prediction:

| Image			        |     Prediction	        					|
|:---------------------:|:---------------------------------------------:|
| Ahead only      		| Ahead only   									|
| Turn right ahead		| Turn right ahead 								|
| 60km/hr				| End of all speed and passing limits			|
| Stop Sign	      		| Stop Sign 					 				|
| 30km/h     			| 30km/h             							|

The model was able to correctly guess 5 of the 5 traffic signs, which gives an accuracy of 100%. This compares favorably to the accuracy on the test set of 95.3

The code for making predictions on my final model is located in the 14th cell of the Ipython notebook.

For the first image, the model is relatively sure that this is a stop sign (probability of 0.6), and the image does contain a stop sign. The top five soft max probabilities were

| Probability         	|     Prediction	        					|
|:---------------------:|:---------------------------------------------:|
| 1.0         			| Ahead only     								|
| 1.0     				| Turn right ahead								|
| 0.93					| 60km/hr										|
| 0.95	      			| Stop Sign 					 				|
| 0.84				    | 30km/h      				        			|